optimse the build-contract-transaction
