+++ cmd/bytomcli/commands/transaction.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/bytom/blockchain"
 	"github.com/bytom/blockchain/txbuilder"
+	"github.com/bytom/contract"
 	"github.com/bytom/util"
 )
 
@@ -21,6 +22,13 @@ func init() {
 	buildTransactionCmd.PersistentFlags().StringVarP(&btmGas, "gas", "g", "20000000", "program of receiver")
 	buildTransactionCmd.PersistentFlags().BoolVar(&pretty, "pretty", false, "pretty print json result")
 	buildTransactionCmd.PersistentFlags().BoolVar(&alias, "alias", false, "use alias build transaction")
+
+	buildContractTransactionCmd.PersistentFlags().StringVarP(&btmGas, "gas", "g", "20000000", "program of receiver")
+	buildContractTransactionCmd.PersistentFlags().BoolVar(&pretty, "pretty", false, "pretty print json result")
+	buildContractTransactionCmd.PersistentFlags().BoolVar(&alias, "alias", false, "use alias build transaction")
+	buildContractTransactionCmd.PersistentFlags().StringVarP(&contractName, "contract-name", "c", "",
+		"name of template contract, currently supported: 'LockWithPublicKey', 'LockWithMultiSig', 'LockWithPublicKeyHash',"+
+			"\n\t\t\t       'RevealPreimage', 'TradeOffer', 'Escrow', 'CallOption', 'LoanCollateral'")
 
 	signTransactionCmd.PersistentFlags().StringVarP(&password, "password", "p", "", "password of the account which sign these transaction(s)")
 	signTransactionCmd.PersistentFlags().BoolVar(&pretty, "pretty", false, "pretty print json result")
@@ -44,6 +52,7 @@ var (
 	txID            = ""
 	account         = ""
 	detail          = false
+	contractName    = ""
 )
 
 var buildIssueReqFmt = `
@@ -196,6 +205,85 @@ var buildTransactionCmd = &cobra.Command{
 		}
 
 		jww.FEEDBACK.Printf("Template Type: %s\n%s\n", buildType, string(rawTemplate))
+	},
+}
+
+var buildContractTransactionCmd = &cobra.Command{
+	Use:   "build-contract-transaction <outputID> <accountID|alias> <assetID|alias> <amount> -c <contractName> <contractArgs>",
+	Short: "Build transaction for template contract, default use account id and asset id",
+	Args:  cobra.RangeArgs(4, 20),
+	PreRun: func(cmd *cobra.Command, args []string) {
+		cmd.MarkFlagRequired("contract-name")
+	},
+	Run: func(cmd *cobra.Command, args []string) {
+		var buildReqStr string
+		var err error
+
+		minArgsCount := 4
+		usage := "Usage:\n  bytomcli build-contract-transaction <outputID> <accountID|alias> <assetID|alias> <amount> -c <contractName>"
+
+		if ok := contract.CheckContractArgs(contractName, args, minArgsCount, usage); !ok {
+			os.Exit(util.ErrLocalExe)
+		}
+
+		buildReqStr, err = contract.BuildContractTransaction(contractName, args, minArgsCount, alias, btmGas)
+		if err != nil {
+			jww.ERROR.Println(err)
+			os.Exit(util.ErrLocalExe)
+		}
+		fmt.Println("buildReqStr", buildReqStr)
+
+		var buildReq blockchain.BuildRequest
+		if err := json.Unmarshal([]byte(buildReqStr), &buildReq); err != nil {
+			jww.ERROR.Println(err)
+			os.Exit(util.ErrLocalExe)
+		}
+
+		data, exitCode := util.ClientCall("/build-transaction", &buildReq)
+		if exitCode != util.Success {
+			os.Exit(exitCode)
+		}
+
+		dataMap, ok := data.(map[string]interface{})
+		if ok != true {
+			jww.ERROR.Println("invalid type assertion")
+			os.Exit(util.ErrLocalParse)
+		}
+
+		rawTemplate, err := json.Marshal(dataMap)
+		if err != nil {
+			jww.ERROR.Println(err)
+			os.Exit(util.ErrLocalParse)
+		}
+
+		/********************add arguments for contract*********************/
+		var tpl *txbuilder.Template
+		err = json.Unmarshal(rawTemplate, &tpl)
+		if err != nil {
+			jww.ERROR.Println(err)
+			os.Exit(util.ErrLocalParse)
+		}
+
+		var contractArgs []string
+		count := minArgsCount
+		for count < len(args) {
+			contractArgs = append(contractArgs, args[count])
+			count++
+		}
+
+		tpl, err = contract.AddContractArguments(tpl, contractName, contractArgs)
+		if err != nil {
+			jww.ERROR.Println(err)
+			os.Exit(util.ErrLocalParse)
+		}
+
+		addWitnessTemplate, err := json.Marshal(tpl)
+		if err != nil {
+			jww.ERROR.Println(err)
+			os.Exit(util.ErrLocalParse)
+		}
+
+		jww.FEEDBACK.Printf("\ntxbuilder.Template: \n%s\n", string(addWitnessTemplate))
 	},
 }
